+++ src/index.ts
@@ -8,7 +8,11 @@ export {
 } from "./locate";
 export { delegateClick, delegateEvent } from "./delegateClick";
 export { clickOutside } from "./clickOutside";
-export { calculatePosition, getAbsParent } from "./positioning";
+export {
+	calculatePosition,
+	getAbsParent,
+	getPopupParents,
+} from "./positioning";
 export { uid } from "./common";
 export { hotkeys } from "./hotkeys";
 export { dateToString, getDuodecade } from "./dates";